backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
requests
